Suppress Read() error on closing websocket-proxy connection
When one end of the websocket proxy closes the connection (e.g.
a browser reload) a net.OpError 'use of closed network connection'
error will be triggered in the blocking Read()/ReadMessage() call
of the other direction. This is a rather hacky attempt to detect
this situation and not trigger a log.Warning().

Signed-off-by: Rudolph Bott <r@spam.wtf>