use worker option for websocket upgrade
Since Apache 2.4.47 the ProxyPass worker supports the option
upgrade=websocket[1]. Since I was unsuccessful to proxy the websocket
for the console with the supplied Rewrite-method (the connection didn't
upgrade and the ws client received HTTP status 200), I tried the
built-in option with success. Maybe this is the way to go?

[1] https://httpd.apache.org/docs/2.4/mod/mod_proxy.html#wsupgrade

Signed-off-by: Sascha Lucas <sascha_lucas@web.de>